+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 5.3-SNAPSHOT
 
 #### Bugs
+* Fix #2620: Add support for `config.openshift.io/v1` Ingress
 
 - Fix: CRD generator no longer treat enum values as properties (performance)